admiralvaccine: Vaccine Extension Package for ADaM in 'R' Asset Library

Programming vaccine specific Clinical Data Interchange Standards Consortium (CDISC) compliant Analysis Data Model (ADaM) datasets in 'R'. Flat model is followed as per Center for Biologics Evaluation and Research (CBER) guidelines for creating vaccine specific domains. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team (2021), https://www.cdisc.org/standards/foundational/adam/adamig-v1-3-release-package). The package is an extension package of the 'admiral' package.

See also

Author

Maintainer: Arjun Rubalingam arjun.rubalingam@pfizer.com

Authors:

Sukalpo Saha sukalpo.saha@pfizer.com [contributor]

Ben Straub ben.x.straub@gsk.com

Vikram S

Dhivya Kanagaraj

Federico Baratin

Yamini Purna Bollu

Ilse Augustyns

Kalyani Bodicherla

Hilde Delanghe

Lee Armishaw

Other contributors:

Neetu Sangari [contributor]

Abdul Khayat [contributor]

Ankur Jindal [contributor]

Jayashree V [contributor]

Jagadish Katam [contributor]

Andrea Pammolli [contributor]

Daniele Bottigliengo [contributor]

Ranya Ben Hsain [contributor]

Marleen Nijs [contributor]

Mandy Peng [contributor]

Tina Zhai [contributor]

Ross Farrugia [contributor]

Stefan Bundfuss [contributor]

Edoardo Mancini [contributor]

Pfizer Inc. [copyright holder, funder]

GlaxoSmithKline LLC [copyright holder, funder]

Johnson & Johnson [copyright holder, funder]