Creating Severity Records from Diameters

To derive the severity records from the diameter records per subject per event per period.

Usage

derive_param_diam_to_sev(
  dataset = NULL,
  filter_diam = "DIAMETER",
  filter_faobj = c("REDNESS", "SWELLING"),
  testcd_sev = "SEV",
  test_sev = "Severity/Intensity",
  none = c(0, 2),
  mild = c(2, 5),
  mod = c(5, 10),
  sev = 10
)

Arguments

dataset

Input data set

The variables USUBJID, APTPTREF, FAOBJ, FASCAT,AVAL, AVALC, FAOBJ, FATESTCD and FATEST are expected for Input data set.(data set)

filter_diam

Diameter record filter

Default: "DIAMETER" Permitted Value: A character vector or scalar.

Helps to filter the diameter records to derive the severity records by passing the FATESTCD value for diameter which is corresponding to the specified events in filter_faobj.

filter_faobj

Event filter

Default: c("REDNESS","SWELLING") Permitted Value: A character vector or Scalar.

Helps to filter the events (Redness and swelling) which has diameter records to derive severity records by passing the events from FAOBJ.

testcd_sev

FATESTCDvalue for severity

Default: "SEV" Permitted Value: A Character scalar

Assign the value for FATESTCD variable to indicate the severity records. Ignore the argument if you want to set the default value.

Note: This argument value will be used to check whether input data set has Severity records for specified filter_faobj event. If it has, those records will be removed and new severity records will be derived from diameters.

test_sev

FATEST value for severity

Default: "Severity" Permitted Value: A Character scalar

Assign the value for FATEST variable to indicate the severity records. Ignore the argument if you want to set the default value.

Assign the Severity limits to create Severity grade(AVALC)

none

limit for grade "NONE"

Default: c(0,2) Permitted value: A Numeric vector

The none and the following arguments(mild, mode and sev ) will be used for assigning the diameter limits to derive the AVALC(severity grade).

Note: Use the limit reference to pass the values in argument Since the condition was coded like this, NONE : none < AVAL <= none MILD : mild < AVAL <= mild MODERATE: mod < AVAL <= mod SEVERE: sev > AVAL User should pass the values as numeric vectors. Refer the default values.

mild

limit for grade "MILD"

Default: c(2,5) Permitted value: A Numeric vector

mod

limit for grade "MODERATE"

Default: c(5,10) Permitted value: A Numeric vector

sev

limit for grade "SEVERE"

Default: 10 Permitted value: A Numeric vector

Value

The Input data with the new severity records for Redness and swelling which is specified in filter_faobj and AVAL, AVALC will be derived and fatestcd, fatest will be changed as per the values

Details

Pass the Input data set in dataset with required variables and AVAL should have the diameter values for the events(FAOBJ) specified in filter_faobj.

If the SDTM data set has severity record for redness and swelling, User can skip this function to keep the SDTM level severity records for further derivation.

If User want to derive the severity records from the diameter records even that data set has severity records for Redness and swelling. This function will remove the existing severity records and it will create the new severity records asAVALC which is derived from diameter. and AVAL will be derived as numeric severity grade to get the maximum severity records which will be derived by derive_param_maxsev.R.

Pass the values in severity grade arguments(none, mild, mod, sev) as per the study needs. 5.User can pass the events which has diameter values to derive their severity records by passing the appropriate filter_diam, filter_faobjand severity grade limits.

See also

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(admiral)
library(tibble)

input <- tribble(
  ~USUBJID, ~FAOBJ, ~AVAL, ~AVALC, ~ATPTREF, ~FATEST, ~FATESTCD,
  "XYZ1001", "REDNESS", 7.5, "7.5", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1001", "REDNESS", 3.5, "3.5", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1001", "REDNESS", 2, "2", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1001", "REDNESS", 1.8, "1.8", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1001", "REDNESS", 1.4, "1.4", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1002", "REDNESS", 11.1, "11.1", "VACCINATION 2", "Diameter", "DIAMETER",
  "XYZ1002", "REDNESS", 7.4, "7.4", "VACCINATION 2", "Diameter", "DIAMETER",
  "XYZ1002", "REDNESS", 6, "6", "VACCINATION 2", "Diameter", "DIAMETER",
  "XYZ1002", "REDNESS", 2.1, "2.1", "VACCINATION 2", "Diameter", "DIAMETER",
  "XYZ1002", "REDNESS", 1.1, "1.1", "VACCINATION 2", "Diameter", "DIAMETER",
  "XYZ1001", "SWELLING", 5.5, "5.5", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1001", "SWELLING", 2.5, "2.5", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1001", "SWELLING", 2, "2", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1001", "SWELLING", 1.8, "1.8", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1001", "SWELLING", 1.4, "1.4", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1002", "SWELLING", 10.1, "10.1", "VACCINATION 2", "Diameter", "DIAMETER",
  "XYZ1002", "SWELLING", 7.1, "7.1", "VACCINATION 2", "Diameter", "DIAMETER",
  "XYZ1002", "SWELLING", 5, "5", "VACCINATION 2", "Diameter", "DIAMETER",
  "XYZ1002", "SWELLING", 1.8, "1.8", "VACCINATION 2", "Diameter", "DIAMETER",
  "XYZ1002", "SWELLING", 1.4, "1.4", "VACCINATION 2", "Diameter", "DIAMETER"
)

derive_param_diam_to_sev(
  dataset = input,
  filter_faobj = c("REDENSS", "SWELLING"),
  filter_diam = "DIAMETER",
  testcd_sev = "SEV",
  test_sev = "Severity"
)