Adds records for the maximum diameter values within each by group to the output dataset.

Creates maximum diameter record per subject per event per period. Maximum diameter records will be derived for the Administration and systemic events which has diameter records.

Usage

derive_param_maxdiam(
  dataset,
  filter = NULL,
  by_vars = NULL,
  test_maxdiam = "Maximum Diameter",
  testcd_maxdiam = "MAXDIAM"
)

Arguments

dataset

Input dataset

The variables USUBJID, APTPTREF, FAOBJ, FASCAT, AVAL, FATESTCD and FATEST are required for the input data set.(dataset)

filter

Filter condition

The specified condition is applied to the input dataset before deriving the new parameter, i.e., only observations fulfilling the condition are taken into account.

FATESTCD and FAOBJ related to diameter records are expected to be filtered in this step

by_vars

Grouping variables

Default: exprs(USUBJID, FAOBJ, ATPTREF)

Variables to consider for generation of groupwise maximum diameter records. Providing the names of variables in exprs() will create a groupwise summary and generate maximum diam records for the specified groups.

test_maxdiam

Value for FATEST

Default: "Maximum Diameter" Permitted Value: A Character scalar

Pass a value for FATEST which will be indicate the maximum diameter records

testcd_maxdiam

Value for FATESTCD

Default: "MAXDIAM" Permitted Value: A Character scalar

Pass a value for FATESTCD which will be indicate the maximum diameter records.

Value

A data frame with derived records for maximum diameter appended to the original dataset.

Details

The maximum aval value of diameter within each group is filtered.

if more than one record have same maximum values within by group then this function will take the first record from that group.

See also

Author

Vikram S

Examples

library(tibble)
library(admiral)
library(dplyr)
library(rlang)
#> 
#> Attaching package: ‘rlang’
#> The following object is masked from ‘package:admiraldev’:
#> 
#>     is_named

input <- tribble(
  ~USUBJID, ~FAOBJ, ~ATPTREF, ~AVAL, ~AVALC, ~FATEST, ~FATESTCD, ~FASCAT,
  "1", "REDNESS", "VAC1", 3.5, "3.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "1", "REDNESS", "VAC1", 4.5, "4.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "1", "REDNESS", "VAC1", 1.5, "3.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "1", "REDNESS", "VAC2", 3.5, "3.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "1", "REDNESS", "VAC2", 4.5, "4.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "1", "REDNESS", "VAC2", 1.5, "3.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "2", "SWELLING", "VAC1", 3.5, "3.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "2", "SWELLING", "VAC1", 4.5, "4.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "2", "SWELLING", "VAC1", 1.5, "3.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "2", "SWELLING", "VAC2", .5, "3.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "2", "SWELLING", "VAC2", 4.5, "4.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "2", "SWELLING", "VAC2", 1.5, "3.5", "Diameter", "DIAMETER", "ADMIN-SITE"
)

# Derive maximum diameter records for redness and swelling.

derive_param_maxdiam(
  dataset = input,
  filter = FAOBJ %in% c("REDNESS", "SWELLING") & FATESTCD == "DIAMETER",
  by_vars = exprs(USUBJID, FAOBJ, ATPTREF),
  test_maxdiam = "Maximum Diameter",
  testcd_maxdiam = "MAXDIAM"
)
#> # A tibble: 16 × 9
#>    USUBJID FAOBJ    ATPTREF  AVAL AVALC FATEST           FATESTCD FASCAT   DTYPE
#>    <chr>   <chr>    <chr>   <dbl> <chr> <chr>            <chr>    <chr>    <chr>
#>  1 1       REDNESS  VAC1      3.5 3.5   Diameter         DIAMETER ADMIN-S… NA   
#>  2 1       REDNESS  VAC1      4.5 4.5   Diameter         DIAMETER ADMIN-S… NA   
#>  3 1       REDNESS  VAC1      1.5 3.5   Diameter         DIAMETER ADMIN-S… NA   
#>  4 1       REDNESS  VAC2      3.5 3.5   Diameter         DIAMETER ADMIN-S… NA   
#>  5 1       REDNESS  VAC2      4.5 4.5   Diameter         DIAMETER ADMIN-S… NA   
#>  6 1       REDNESS  VAC2      1.5 3.5   Diameter         DIAMETER ADMIN-S… NA   
#>  7 2       SWELLING VAC1      3.5 3.5   Diameter         DIAMETER ADMIN-S… NA   
#>  8 2       SWELLING VAC1      4.5 4.5   Diameter         DIAMETER ADMIN-S… NA   
#>  9 2       SWELLING VAC1      1.5 3.5   Diameter         DIAMETER ADMIN-S… NA   
#> 10 2       SWELLING VAC2      0.5 3.5   Diameter         DIAMETER ADMIN-S… NA   
#> 11 2       SWELLING VAC2      4.5 4.5   Diameter         DIAMETER ADMIN-S… NA   
#> 12 2       SWELLING VAC2      1.5 3.5   Diameter         DIAMETER ADMIN-S… NA   
#> 13 1       REDNESS  VAC1      4.5 NA    Maximum Diameter MAXDIAM  ADMIN-S… MAXI…
#> 14 1       REDNESS  VAC2      4.5 NA    Maximum Diameter MAXDIAM  ADMIN-S… MAXI…
#> 15 2       SWELLING VAC1      4.5 NA    Maximum Diameter MAXDIAM  ADMIN-S… MAXI…
#> 16 2       SWELLING VAC2      4.5 NA    Maximum Diameter MAXDIAM  ADMIN-S… MAXI…