Creating Maximum Severity Records

Creating the maximum severity records per subject per event per period. Maximum severity records will be derived for the Administration and systemic events which has severity records FATESTCD = "SEV".

Usage

derive_param_maxsev(
  dataset,
  exclude_events = NULL,
  filter_sev = "SEV",
  test_maxsev = "Maximum Severity",
  testcd_maxsev = "MAXSEV",
  by_vars = exprs(USUBJID, FAOBJ, ATPTREF)
)

Arguments

dataset

input data set

The variables USUBJID, APTPTREF, FAOBJ, FASCAT, AVALC, FAOBJ, FATESTCD and FATEST are required for the input data set.(dataset)

exclude_events

To exclude the events

Helps to exclude the events which user doesn't want to derive the maximum severity. Pass the event name in quotes which is from FAOBJ.

Default: NULL Permitted Value: A character vector.

filter_sev

severity record filter

Default: "SEV" Permitted Value: A character scalar

To filter the severity records to derive the maximum severity by passing the FATESTCD value for severity in quotes "SEV" for this argument.

test_maxsev

FATEST value for maximum severity records

Default: "Maximum severity" Permitted Value: A character scalar

Assign the value for FATEST variable to indicate the maximum severity records. Ignore the argument if you want to set the value as mentioned in default.

testcd_maxsev

FATESTCD value for maximum severity records

Default: "MAXSEV" Permitted Value: A character scalar

Assign the value for FATESTCD variable to indicate the maximum severity records. Ignore the argument if you want to set the value as mentioned in default.

by_vars

Grouping variables

Default: exprs(USUBJID, FAOBJ, ATPTREF)

Based on the variables which will be passed on this by_vars argument, The maximum severity records will be derived from AVAL. Note: Pass the variables in exprs().

Value

The input data set with new records FATESTCD = MAXSEVindicating the maximum severity records for the specified variables in by_vars. DTYPE

will be populated as MAXIMUM. FATESTwill be populated as specified in test_maxsevand FATESCD will be populated as specified in testcd_maxsev.

See also

Author

Arjun Rubalingam

Examples

library(tibble)
library(admiral)
library(dplyr)
library(rlang)

input <- tribble(
  ~USUBJID, ~FAOBJ, ~AVAL, ~AVALC, ~ATPTREF, ~FATEST, ~FATESTCD, ~FASCAT,
  "XYZ1001", "REDNESS", 1, "MILD", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "REDNESS", 2, "MODERATE", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "REDNESS", 2, "MODERATE", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "REDNESS", 2, "MODERATE", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "REDNESS", 3, "SEVERE", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "REDNESS", 2, "MILD", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "REDNESS", 2, "MODERATE", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "REDNESS", 1, "MILD", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "REDNESS", 1, "MILD", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "REDNESS", 0, "NONE", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 1, "MILD", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 1, "MILD", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 0, "NONE", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 1, "MILD", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 0, "NONE", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 2, "MODERATE", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 1, "MILD", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 0, "NONE", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 3, "SEVERE", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 0, "NONE", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "CHILLS", NA, "MODERATE", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1001", "CHILLS", NA, "MILD", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1001", "CHILLS", NA, "NONE", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1001", "CHILLS", NA, "MILD", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1001", "CHILLS", NA, "NONE", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1001", "CHILLS", NA, "MODERATE", "VACC2", "Severity", "SEV", "SYSTEMIC",
  "XYZ1001", "CHILLS", NA, "MILD", "VACC2", "Severity", "SEV", "SYSTEMIC",
  "XYZ1001", "CHILLS", NA, "NONE", "VACC2", "Severity", "SEV", "SYSTEMIC",
  "XYZ1001", "CHILLS", NA, "MILD", "VACC2", "Severity", "SEV", "SYSTEMIC",
  "XYZ1001", "CHILLS", NA, "NONE", "VACC2", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "MODERATE", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "MILD", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "NONE", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "NONE", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "NONE", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "MILD", "VACC2", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "NONE", "VACC2", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "NONE", "VACC2", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "MODERATE", "VACC2", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "NONE", "VACC2", "Severity", "SEV", "SYSTEMIC"
)

derive_param_maxsev(
  dataset = input,
  filter_sev = "SEV",
  exclude_events = "REDNESS",
  test_maxsev = "Maximum severity",
  testcd_maxsev = "MAXSEV",
  by_vars = exprs(USUBJID, FAOBJ, ATPTREF)
)
#>    USUBJID   FAOBJ AVAL    AVALC ATPTREF           FATEST FATESTCD     FASCAT
#> 1  XYZ1001 REDNESS    1     MILD   VACC1         Severity      SEV ADMIN-SITE
#> 2  XYZ1001 REDNESS    2 MODERATE   VACC1         Severity      SEV ADMIN-SITE
#> 3  XYZ1001 REDNESS    2 MODERATE   VACC1         Severity      SEV ADMIN-SITE
#> 4  XYZ1001 REDNESS    2 MODERATE   VACC1         Severity      SEV ADMIN-SITE
#> 5  XYZ1001 REDNESS    3   SEVERE   VACC1         Severity      SEV ADMIN-SITE
#> 6  XYZ1001 REDNESS    1     MILD   VACC2         Severity      SEV ADMIN-SITE
#> 7  XYZ1001 REDNESS    2 MODERATE   VACC2         Severity      SEV ADMIN-SITE
#> 8  XYZ1001 REDNESS    1     MILD   VACC2         Severity      SEV ADMIN-SITE
#> 9  XYZ1001 REDNESS    1     MILD   VACC2         Severity      SEV ADMIN-SITE
#> 10 XYZ1001 REDNESS    0     NONE   VACC2         Severity      SEV ADMIN-SITE
#> 11 XYZ1002 REDNESS    1     MILD   VACC1         Severity      SEV ADMIN-SITE
#> 12 XYZ1002 REDNESS    1     MILD   VACC1         Severity      SEV ADMIN-SITE
#> 13 XYZ1002 REDNESS    0     NONE   VACC1         Severity      SEV ADMIN-SITE
#> 14 XYZ1002 REDNESS    1     MILD   VACC1         Severity      SEV ADMIN-SITE
#> 15 XYZ1002 REDNESS    0     NONE   VACC1         Severity      SEV ADMIN-SITE
#> 16 XYZ1002 REDNESS    2 MODERATE   VACC2         Severity      SEV ADMIN-SITE
#> 17 XYZ1002 REDNESS    1     MILD   VACC2         Severity      SEV ADMIN-SITE
#> 18 XYZ1002 REDNESS    0     NONE   VACC2         Severity      SEV ADMIN-SITE
#> 19 XYZ1002 REDNESS    3   SEVERE   VACC2         Severity      SEV ADMIN-SITE
#> 20 XYZ1002 REDNESS    0     NONE   VACC2         Severity      SEV ADMIN-SITE
#> 21 XYZ1001  CHILLS    2 MODERATE   VACC1         Severity      SEV   SYSTEMIC
#> 22 XYZ1001  CHILLS    1     MILD   VACC1         Severity      SEV   SYSTEMIC
#> 23 XYZ1001  CHILLS    0     NONE   VACC1         Severity      SEV   SYSTEMIC
#> 24 XYZ1001  CHILLS    1     MILD   VACC1         Severity      SEV   SYSTEMIC
#> 25 XYZ1001  CHILLS    0     NONE   VACC1         Severity      SEV   SYSTEMIC
#> 26 XYZ1001  CHILLS    2 MODERATE   VACC2         Severity      SEV   SYSTEMIC
#> 27 XYZ1001  CHILLS    1     MILD   VACC2         Severity      SEV   SYSTEMIC
#> 28 XYZ1001  CHILLS    0     NONE   VACC2         Severity      SEV   SYSTEMIC
#> 29 XYZ1001  CHILLS    1     MILD   VACC2         Severity      SEV   SYSTEMIC
#> 30 XYZ1001  CHILLS    0     NONE   VACC2         Severity      SEV   SYSTEMIC
#> 31 XYZ1002  CHILLS    2 MODERATE   VACC1         Severity      SEV   SYSTEMIC
#> 32 XYZ1002  CHILLS    1     MILD   VACC1         Severity      SEV   SYSTEMIC
#> 33 XYZ1002  CHILLS    0     NONE   VACC1         Severity      SEV   SYSTEMIC
#> 34 XYZ1002  CHILLS    0     NONE   VACC1         Severity      SEV   SYSTEMIC
#> 35 XYZ1002  CHILLS    0     NONE   VACC1         Severity      SEV   SYSTEMIC
#> 36 XYZ1002  CHILLS    1     MILD   VACC2         Severity      SEV   SYSTEMIC
#> 37 XYZ1002  CHILLS    0     NONE   VACC2         Severity      SEV   SYSTEMIC
#> 38 XYZ1002  CHILLS    0     NONE   VACC2         Severity      SEV   SYSTEMIC
#> 39 XYZ1002  CHILLS    2 MODERATE   VACC2         Severity      SEV   SYSTEMIC
#> 40 XYZ1002  CHILLS    0     NONE   VACC2         Severity      SEV   SYSTEMIC
#> 41 XYZ1001  CHILLS    2 MODERATE   VACC1 Maximum severity   MAXSEV   SYSTEMIC
#> 42 XYZ1001  CHILLS    2 MODERATE   VACC2 Maximum severity   MAXSEV   SYSTEMIC
#> 43 XYZ1002  CHILLS    2 MODERATE   VACC1 Maximum severity   MAXSEV   SYSTEMIC
#> 44 XYZ1002  CHILLS    2 MODERATE   VACC2 Maximum severity   MAXSEV   SYSTEMIC
#>      DTYPE
#> 1     <NA>
#> 2     <NA>
#> 3     <NA>
#> 4     <NA>
#> 5     <NA>
#> 6     <NA>
#> 7     <NA>
#> 8     <NA>
#> 9     <NA>
#> 10    <NA>
#> 11    <NA>
#> 12    <NA>
#> 13    <NA>
#> 14    <NA>
#> 15    <NA>
#> 16    <NA>
#> 17    <NA>
#> 18    <NA>
#> 19    <NA>
#> 20    <NA>
#> 21    <NA>
#> 22    <NA>
#> 23    <NA>
#> 24    <NA>
#> 25    <NA>
#> 26    <NA>
#> 27    <NA>
#> 28    <NA>
#> 29    <NA>
#> 30    <NA>
#> 31    <NA>
#> 32    <NA>
#> 33    <NA>
#> 34    <NA>
#> 35    <NA>
#> 36    <NA>
#> 37    <NA>
#> 38    <NA>
#> 39    <NA>
#> 40    <NA>
#> 41 MAXIMUM
#> 42 MAXIMUM
#> 43 MAXIMUM
#> 44 MAXIMUM