Creating Maximum Severity Records

Creating the maximum severity records per subject per event per period. Maximum severity records will be derived for the Administration and systemic events which has severity records FATESTCD = "SEV".

Usage

derive_param_maxsev(
  dataset = NULL,
  exclude_events = NULL,
  filter_sev = "SEV",
  test_maxsev = "Maximum Severity",
  testcd_maxsev = "MAXSEV",
  by_vars = exprs(USUBJID, FAOBJ, ATPTREF)
)

Arguments

dataset

input data set

The variables USUBJID, APTPTREF, FAOBJ, FASCAT, AVALC, FAOBJ, FATESTCD and FATEST are required for the input data set.(dataset)

exclude_events

To exclude the events

Helps to exclude the events which user doesn't want to derive the maximum severity. Pass the event name in quotes which is from FAOBJ.

Default: NULL Permitted Value: A character vector.

filter_sev

severity record filter

Default: "SEV" Permitted Value: A character scalar

To filter the severity records to derive the maximum severity by passing the FATESTCD value for severity in quotes "SEV" for this argument.

test_maxsev

FATEST value for maximum severity records

Default: "Maximum severity" Permitted Value: A character scalar

Assign the value for FATEST variable to indicate the maximum severity records. Ignore the argument if you want to set the value as mentioned in default.

testcd_maxsev

FATESTCD value for maximum severity records

Default: "MAXSEV" Permitted Value: A character scalar

Assign the value for FATESTCD variable to indicate the maximum severity records. Ignore the argument if you want to set the value as mentioned in default.

by_vars

Grouping variables

Default: exprs(USUBJID, FAOBJ, ATPTREF)

Based on the variables which will be passed on this by_vars argument, The maximum severity records will be derived from AVAL. Note: Pass the variables in exprs().

Value

The input data set with new records FATESTCD = MAXSEVindicating the maximum severity records for the specified variables in by_vars. DTYPE

will be populated as MAXIMUM. FATESTwill be populated as specified in test_maxsevand FATESCD will be populated as specified in testcd_maxsev.

See also

Author

Arjun Rubalingam

Examples

library(tibble)
library(admiral)
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following object is masked from ‘package:admiral’:
#> 
#>     vars
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(rlang)

input <- tribble(
  ~USUBJID, ~FAOBJ, ~AVAL, ~AVALC, ~ATPTREF, ~FATEST, ~FATESTCD, ~FASCAT,
  "XYZ1001", "REDNESS", 1, "MILD", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "REDNESS", 2, "MODERATE", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "REDNESS", 2, "MODERATE", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "REDNESS", 2, "MODERATE", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "REDNESS", 3, "SEVERE", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "REDNESS", 2, "MILD", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "REDNESS", 2, "MODERATE", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "REDNESS", 1, "MILD", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "REDNESS", 1, "MILD", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "REDNESS", 0, "NONE", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 1, "MILD", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 1, "MILD", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 0, "NONE", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 1, "MILD", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 0, "NONE", "VACC1", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 2, "MODERATE", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 1, "MILD", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 0, "NONE", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 3, "SEVERE", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1002", "REDNESS", 0, "NONE", "VACC2", "Severity", "SEV", "ADMIN-SITE",
  "XYZ1001", "CHILLS", NA, "MODERATE", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1001", "CHILLS", NA, "MILD", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1001", "CHILLS", NA, "NONE", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1001", "CHILLS", NA, "MILD", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1001", "CHILLS", NA, "NONE", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1001", "CHILLS", NA, "MODERATE", "VACC2", "Severity", "SEV", "SYSTEMIC",
  "XYZ1001", "CHILLS", NA, "MILD", "VACC2", "Severity", "SEV", "SYSTEMIC",
  "XYZ1001", "CHILLS", NA, "NONE", "VACC2", "Severity", "SEV", "SYSTEMIC",
  "XYZ1001", "CHILLS", NA, "MILD", "VACC2", "Severity", "SEV", "SYSTEMIC",
  "XYZ1001", "CHILLS", NA, "NONE", "VACC2", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "MODERATE", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "MILD", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "NONE", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "NONE", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "NONE", "VACC1", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "MILD", "VACC2", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "NONE", "VACC2", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "NONE", "VACC2", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "MODERATE", "VACC2", "Severity", "SEV", "SYSTEMIC",
  "XYZ1002", "CHILLS", NA, "NONE", "VACC2", "Severity", "SEV", "SYSTEMIC"
)

derive_param_maxsev(
  dataset = input,
  filter_sev = "SEV",
  exclude_events = "REDNESS",
  test_maxsev = "Maximum severity",
  testcd_maxsev = "MAXSEV",
  by_vars = exprs(USUBJID, FAOBJ, ATPTREF)
)
#> Error in assert_list_of(arg, "symbol", named = expect_names, optional = optional): Each element of `arg` must be an object of class/type 'symbol' but the following are not:
#> ✖ Element 1 is an object of class 'quosure'
#> ✖ Element 2 is an object of class 'quosure'
#> ✖ Element 3 is an object of class 'quosure'
#> ✖ Element 4 is an object of class 'quosure'
#> ✖ Element 5 is an object of class 'quosure'
#> ✖ Element 6 is an object of class 'quosure'
#> ✖ Element 7 is an object of class 'quosure'